Working Set

+++ MapReduceDental/src/main/java/group05/bigdata/mappers/StatePopulationMapper.java
@@ -1,11 +1,12 @@
 package group05.bigdata.mappers;
 
 import java.io.IOException;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
-public class StatePopulationMapper extends  Mapper<LongWritable, Text, Text, LongWritable> {
+public class StatePopulationMapper extends  Mapper<LongWritable, Text, Text, FloatWritable> {
 	
 	public void map(LongWritable lineOffSet, Text record, Context context)
 			throws IOException, InterruptedException {
@@ -23,8 +24,8 @@ public class StatePopulationMapper extends  Mapper<LongWritable, Text, Text, Lon
         }	
 		String stateKey = columnValues[0].trim();      
 		try{
-			long population = Long.parseLong(columnValues[2].trim());
-			context.write(new Text(stateKey), new LongWritable(population));
+			float population = Float.parseFloat(columnValues[2].trim());
+			context.write(new Text(stateKey), new FloatWritable(population));
 		}
 		catch(NumberFormatException e){
 			return;
