+++ MapReduceDental/src/main/java/group05/bigdata/reducer/StatePopulationReducer.java
@@ -1,20 +1,20 @@
 package group05.bigdata.reducer;
 
 import java.io.IOException;
-import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
-public class StatePopulationReducer extends Reducer <Text, LongWritable, Text, LongWritable>{
+public class StatePopulationReducer extends Reducer <Text, FloatWritable, Text, FloatWritable>{
 	
-	public void reduce(Text key, Iterable<LongWritable> values,
+	public void reduce(Text key, Iterable<FloatWritable> values,
             Context context) throws IOException, InterruptedException {
 		long population = 0;
         // Calculate the total population in case of multiple values
-       for (LongWritable value : values) {
+       for (FloatWritable value : values) {
     	   population += value.get();
         } 
-        context.write(key, new LongWritable(population));
+        context.write(key, new FloatWritable(population));
         // Key is State, Value is the Population  
     }		
 }
